fix(layering): move tokenize_agent_target to models/types/conversations
unirl/models/qwen_vl/chat_template.py importing it from unirl.data.sft (#352)
is a models->data edge check_core_dependencies forbids, and it failed the new
hook on the PR's own merged tree. The function is pure tokenizer mechanics
(record messages -> final-assistant-turn token suffix) with no data-layer
dependency, so it belongs with the other trainside conversation rendering in
models/types/conversations.py, which every consumer (models, train/sft,
datasets CLI) may legally import. No re-export left behind.
