+++ unirl/train/sft/track_builder.py
@@ -8,7 +8,8 @@ from typing import Any, Dict, List, Optional, Protocol, Sequence, Tuple
 
 import torch
 
-from unirl.data.sft import message_content_image_uris, tokenize_agent_target
+from unirl.data.sft import message_content_image_uris
+from unirl.models.types.conversations import tokenize_agent_target
 from unirl.distributed.group.dispatch import Dispatch, distributed
 from unirl.distributed.group.remote import Remote
 from unirl.models.types.codec import EncodeStage
